2.3.0 Changelog:  - Added /cc bypass Bug Fixes:  - Fixed the chat prompts for command labels  - Fixed the Editor GUI not allowing the user to change cooldowns  - Fixed the Editor GUI incorrectly creating commands  - Fixed loading/saving cooldowns when enabling and disabling the plugin  - Fixed cooldowns not applying to players

+++ src/main/java/me/darrionat/commandcooldown/interfaces/ICooldownService.java
@@ -21,13 +21,13 @@ public interface ICooldownService extends Service {
     void giveCooldown(Player p, Cooldown cooldown);
 
     /**
-     * Gets the player's remaining cooldown for a particular command.
+     * Gets the player's remaining cooldown for a particular command in seconds.
      *
      * @param p        the player.
      * @param cooldown the command with a cooldown.
      * @return returns a player's remaining cooldown for the command; {@code -1} if the player has no cooldown.
      */
-    long getRemainingCooldown(Player p, Cooldown cooldown);
+    double getRemainingCooldown(Player p, Cooldown cooldown);
 
     /**
      * Determines if the player has a cooldown for a command.
@@ -43,7 +43,8 @@ public interface ICooldownService extends Service {
      *
      * @param p        The player.
      * @param cooldown The cooldown to check for duration changes.
-     * @return The cooldown with a changed duration if the player's permissions affect it; otherwise the original cooldown.
+     * @return The cooldown with a changed duration if the player's permissions affect it; otherwise the original
+     *         cooldown.
      */
     Cooldown permissionCooldownChange(Player p, Cooldown cooldown);
